Properly initialize new Donut/Gauge object when props.donut changes
Previously, in the case that props.donut was changed after the component
was mounted, gauge.current would be set to a new Donut/Gauge object, but
that object would not be initialized properly.

This fix comes with the disadvantage that everything will be initialized twice
when the component is first mounted.